requirements.txt
# Garmin API
garminconnect>=0.2.37

# Data processing
pandas>=2.0.0
numpy>=2.0.0

# Machine learning
scikit-learn>=1.5.0
scipy>=1.10.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Dashboard
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0

# BigQuery (optional)
google-cloud-bigquery>=3.0.0
db-dtypes>=1.0.0
protobuf>=4.0.0

# Jupyter (optional)
jupyter>=1.0.0
nbconvert>=7.0.0
